componentize search components
